+++ crates/storb_validator/src/constants.rs
@@ -1,0 +1,1 @@
+pub const MIN_REQUIRED_MINERS: usize = 1;
